a debugger source was separated.
