+++ src/allocator.c
@@ -92,6 +92,7 @@ kx_context_t *make_context(void)
     ctx->options.case_threshold = 16;
     ctx->options.max_call_depth = 1024;
     ctx->ir_dumpcode = ir_code_dump_one;
+    ctx->ir_executor = ir_exec;
     return ctx;
 }
